requirements.txt
anthropic>=0.49.0
anyio>=4.8.0
apscheduler>=3.11.0
asyncio>=3.4.3
flask>=3.1.0
fpdf>=1.7.2
gunicorn>=20.1.0
matplotlib>=3.10.1
numpy>=2.2.3
openai>=1.75.0
pandas>=2.2.3
pandas-ta>=0.3.14b0
python-dotenv>=1.1.0
python-telegram-bot==20.7
requests>=2.32.3
rich>=14.0.0
schedule>=1.2.2
scikit-learn>=1.6.1
setuptools>=75.8.2
sqlalchemy>=2.0.40
ta>=0.11.0
telegram>=0.0.1
torch>=2.6.0
tzlocal>=5.3
websockets>=15.0
psycopg2-binary>=2.9.9
TA-Lib==0.4.28
pytz>=2023.3
aiohttp>=3.9.1
yfinance>=0.2.18